requirements.txt
beautifulsoup4>=4.12.3
fake-useragent>=1.5.1
pandas>=2.2.3
python-dotenv>=1.0.1
requests>=2.32.3
selenium>=4.29.0
webdriver-manager>=4.0.2
streamlit==1.43.2
